Rename fetchSource to source in ducks
